final commit

+++ Assets/_Scripts/Configs/Scripts/PlayerConfig.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "PlayerConfig", menuName = "Configs/PlayerConfig")]
 public class PlayerConfig : ScriptableObject
 {
-    public int health; 
-    public int damage; 
+    public float health; 
+    public float damage; 
 
     public float speed;
     public float attackSpeed;
