+++ Assets/_Scripts/Interfaces/IDamagable.cs
@@ -1,4 +1,4 @@
 public interface IDamagable
 {
-    void TakeDamage(int value);
+    void TakeDamage(float value);
 }
